Restructure VSCode extension codebase
This is to further clean up the codebase.
Moved each feature related logics to a
separate folder. Refactored debugger file
names to avoid camel case file names. Further
cleaned up debugger code base. Rename DebugPoint
to BreakPoint. Moved config resolver for debugger
to an individual module, etc.
