+++ tool-plugins/vscode/src/bbe/activator.ts
@@ -19,9 +19,9 @@
 import { commands, window, Uri, ViewColumn, ExtensionContext, WebviewPanel, workspace } from 'vscode';
 import * as path from 'path';
 import { render } from './renderer';
-import { ExtendedLangClient } from '../lang-client';
+import { ExtendedLangClient } from '../client';
+import { BallerinaExtInstance } from '../core';
 import { WebViewRPCHandler } from '../utils';
-import BallerinaExtension from '../core/ballerina-extension';
 
 let examplesPanel: WebviewPanel | undefined;
 
@@ -58,12 +58,14 @@ function showExamples(context: ExtensionContext, langClient: ExtendedLangClient)
         switch (message.command) {
             case 'openExample':
                 const url = JSON.parse(message.url);
-                const ballerinaHome = BallerinaExtension.getBallerinaHome();
+                const ballerinaHome = BallerinaExtInstance.getBallerinaHome();
                 if (ballerinaHome) {
                     const folderPath = path.join(ballerinaHome, 'examples', url);
                     const filePath = path.join(folderPath, `${url.replace(/-/g, '_')}.bal`);
                     workspace.openTextDocument(Uri.file(filePath)).then(doc => {
                         window.showTextDocument(doc);
+                     }, (err: Error) => {
+                         window.showErrorMessage(err.message);
                      }); 
                 }
                 break;
@@ -77,28 +79,26 @@ function showExamples(context: ExtensionContext, langClient: ExtendedLangClient)
 
 export function activate(context: ExtensionContext, langClient: ExtendedLangClient) {
     const examplesListRenderer = commands.registerCommand('ballerina.showExamples', () => {
-        BallerinaExtension.onReady()
+        BallerinaExtInstance.onReady()
         .then(() => {
             const { experimental } = langClient.initializeResult!.capabilities;
             const serverProvidesExamples = experimental && experimental.examplesProvider;
 
             if (!serverProvidesExamples) {
-                BallerinaExtension.showMessageServerMissingCapability();
+                BallerinaExtInstance.showMessageServerMissingCapability();
                 return;
             }
 
             showExamples(context, langClient);
         })
 		.catch((e) => {
-			if (!BallerinaExtension.isValidBallerinaHome()) {
-				BallerinaExtension.showMessageInvalidBallerinaHome();
+			if (!BallerinaExtInstance.isValidBallerinaHome()) {
+				BallerinaExtInstance.showMessageInvalidBallerinaHome();
 			} else {
-				BallerinaExtension.showPluginActivationError();
+				BallerinaExtInstance.showPluginActivationError();
 			}
 		});
     });
     
     context.subscriptions.push(examplesListRenderer);
 }
-
-
