+++ tool-plugins/vscode/src/bbe/index.ts
@@ -1,0 +1,1 @@
+export * from './activator';
