+++ tool-plugins/vscode/src/core/index.ts
@@ -1,0 +1,2 @@
+export * from './extension';
+export * from './messages';
