+++ tool-plugins/vscode/src/debugger/start.ts
@@ -1,0 +1,4 @@
+import { BallerinaDebugSession } from './session';
+import { DebugSession } from 'vscode-debugadapter';
+
+DebugSession.run(BallerinaDebugSession);
