+++ tool-plugins/vscode/src/debugger/types.d.ts
@@ -1,3 +1,5 @@
+
+// Declare types of deps without type declarations.
 declare module 'ps-node' {
     export function lookup(args: any, callback: any) : void;
     export function kill(args: any) : void;
