+++ tool-plugins/vscode/src/diagram/index.ts
@@ -1,0 +1,1 @@
+export * from './activator';
