+++ tool-plugins/vscode/src/utils/rpc/index.ts
@@ -1,0 +1,2 @@
+export * from './model';
+export * from './handler';
