+++ tool-plugins/vscode/src/utils/rpc/model.ts
@@ -1,0 +1,17 @@
+
+export interface WebViewMethodHandler {
+    (args: any[]) : Thenable<any>;
+}
+
+export interface WebViewMethod {
+    methodName: string;
+    handler: WebViewMethodHandler;
+}
+
+export interface WebViewRPCMessage {
+    id?: number;
+    methodName?: any;
+    arguments?: any[];
+    originId?: number;
+    response?: any;
+}
